Reference common interface where possible for DeviceDescriptor
